mongo中的index在创建时转为int32

+++ src/common/index/association.go
@@ -28,36 +28,36 @@ var (
 		{
 			Name: common.CCLogicIndexNamePrefix + "bkObjId_bkInstID",
 			Keys: bson.D{
-				{"bk_obj_id", int32(1)},
-				{"bk_inst_id", int32(1)},
+				{"bk_obj_id", 1},
+				{"bk_inst_id", 1},
 			},
 			Background: true,
 		},
 		{
 			Name:       common.CCLogicUniqueIdxNamePrefix + "id",
-			Keys:       bson.D{{"id", int32(1)}},
+			Keys:       bson.D{{"id", 1}},
 			Unique:     true,
 			Background: true,
 		},
 		{
 			Name: common.CCLogicIndexNamePrefix + "bkAsstObjId_bkAsstInstId",
 			Keys: bson.D{
-				{"bk_asst_obj_id", int32(1)},
-				{"bk_asst_inst_id", int32(1)},
+				{"bk_asst_obj_id", 1},
+				{"bk_asst_inst_id", 1},
 			},
 			Background: true,
 		},
 		{
 			Name:       common.CCLogicIndexNamePrefix + "bkAsstID",
-			Keys:       bson.D{{"bk_asst_id", int32(1)}},
+			Keys:       bson.D{{"bk_asst_id", 1}},
 			Background: true,
 		},
 		{
 			Name: common.CCLogicUniqueIdxNamePrefix + "bkInstID_bkAsstInstID_bkObjAsstID",
 			Keys: bson.D{
-				{common.BKInstIDField, int32(1)},
-				{common.BKAsstInstIDField, int32(1)},
-				{common.AssociationObjAsstIDField, int32(1)},
+				{common.BKInstIDField, 1},
+				{common.BKAsstInstIDField, 1},
+				{common.AssociationObjAsstIDField, 1},
 			},
 			Unique:     true,
 			Background: true,
