+++ src/common/index/collections/business_set.go
@@ -27,7 +27,7 @@ var commBizSetIndexes = []types.Index{
 	{
 		Name: common.CCLogicUniqueIdxNamePrefix + "biz_set_id",
 		Keys: bson.D{{
-			common.BKBizSetIDField, int32(1)},
+			common.BKBizSetIDField, 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -35,7 +35,7 @@ var commBizSetIndexes = []types.Index{
 	{
 		Name: common.CCLogicUniqueIdxNamePrefix + "biz_set_name",
 		Keys: bson.D{{
-			common.BKBizSetNameField, int32(1)},
+			common.BKBizSetNameField, 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -43,9 +43,9 @@ var commBizSetIndexes = []types.Index{
 	{
 		Name: common.CCLogicIndexNamePrefix + "biz_set_id_biz_set_name_owner_id",
 		Keys: bson.D{
-			{common.BKBizSetIDField, int32(1)},
-			{common.BKBizSetNameField, int32(1)},
-			{common.BKOwnerIDField, int32(1)},
+			{common.BKBizSetIDField, 1},
+			{common.BKBizSetNameField, 1},
+			{common.BKOwnerIDField, 1},
 		},
 		Background: true,
 	},
