+++ src/common/index/collections/dynamicgroup.go
@@ -36,7 +36,7 @@ var deprecatedDynamicGroupIndexes = []types.Index{
 	{
 		Name: "id_1",
 		Keys: bson.D{{
-			"id", int32(1)},
+			"id", 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -44,16 +44,16 @@ var deprecatedDynamicGroupIndexes = []types.Index{
 	{
 		Name: "id_1_bk_biz_id_1",
 		Keys: bson.D{
-			{"bk_biz_id", int32(1)},
-			{"id", int32(1)},
+			{"bk_biz_id", 1},
+			{"id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_biz_id_1_name_1",
 		Keys: bson.D{
-			{"bk_biz_id", int32(1)},
-			{"name", int32(1)},
+			{"bk_biz_id", 1},
+			{"name", 1},
 		},
 		Unique:     true,
 		Background: true,
