+++ src/common/index/collections/hostbase.go
@@ -32,8 +32,8 @@ var commHostBaseIndexes = []types.Index{
 	{
 		Name: common.CCLogicUniqueIdxNamePrefix + "bkHostInnerIP_bkCloudID",
 		Keys: bson.D{
-			{common.BKHostInnerIPField, int32(1)},
-			{common.BKCloudIDField, int32(1)},
+			{common.BKHostInnerIPField, 1},
+			{common.BKCloudIDField, 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -45,8 +45,8 @@ var commHostBaseIndexes = []types.Index{
 	{
 		Name: common.CCLogicUniqueIdxNamePrefix + "bkHostInnerIPv6_bkCloudID",
 		Keys: bson.D{
-			{common.BKHostInnerIPv6Field, int32(1)},
-			{common.BKCloudIDField, int32(1)},
+			{common.BKHostInnerIPv6Field, 1},
+			{common.BKCloudIDField, 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -62,29 +62,29 @@ var deprecatedHostBaseIndexes = []types.Index{
 	{
 		Name: "bk_host_name_1",
 		Keys: bson.D{{
-			"bk_host_name", int32(1)},
+			"bk_host_name", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_host_innerip_1",
 		Keys: bson.D{{
-			"bk_host_innerip", int32(1)},
+			"bk_host_innerip", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_host_id_1_bk_supplier_account_1",
 		Keys: bson.D{
-			{"bk_host_id", int32(1)},
-			{"bk_supplier_account", int32(1)},
+			{"bk_host_id", 1},
+			{"bk_supplier_account", 1},
 		},
 		Background: true,
 	},
 	/* 	{
 		Name: "innerIP_platID",
 		Keys: bson.D{{
-			"bk_host_innerip", int32(1)}
+			"bk_host_innerip", 1}
 			"bk_cloud_id":     1,
 		},
 		Background: false,
@@ -92,21 +92,21 @@ var deprecatedHostBaseIndexes = []types.Index{
 	{
 		Name: "bk_supplier_account_1",
 		Keys: bson.D{{
-			"bk_supplier_account", int32(1)},
+			"bk_supplier_account", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_cloud_id_1",
 		Keys: bson.D{{
-			"bk_cloud_id", int32(1)},
+			"bk_cloud_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "idx_unique_hostID",
 		Keys: bson.D{{
-			"bk_host_id", int32(1)},
+			"bk_host_id", 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -114,21 +114,21 @@ var deprecatedHostBaseIndexes = []types.Index{
 	{
 		Name: "cloudInstID",
 		Keys: bson.D{{
-			"bk_cloud_inst_id", int32(1)},
+			"bk_cloud_inst_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_idx_bk_asset_id",
 		Keys: bson.D{{
-			"bk_asset_id", int32(1)},
+			"bk_asset_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_os_type_1",
 		Keys: bson.D{{
-			"bk_os_type", int32(1)},
+			"bk_os_type", 1},
 		},
 		Background: true,
 	},
