+++ src/common/index/collections/modulebase.go
@@ -36,57 +36,57 @@ var deprecatedModuleBaseIndexes = []types.Index{
 	{
 		Name: "bk_module_name_1",
 		Keys: bson.D{{
-			"bk_module_name", int32(1)},
+			"bk_module_name", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "default_1",
 		Keys: bson.D{{
-			"default", int32(1)},
+			"default", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_biz_id_1",
 		Keys: bson.D{{
-			"bk_biz_id", int32(1)},
+			"bk_biz_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_supplier_account_1",
 		Keys: bson.D{{
-			"bk_supplier_account", int32(1)},
+			"bk_supplier_account", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_set_id_1",
 		Keys: bson.D{{
-			"bk_set_id", int32(1)},
+			"bk_set_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_parent_id_1",
 		Keys: bson.D{{
-			"bk_parent_id", int32(1)},
+			"bk_parent_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_module_id_1_bk_biz_id_1",
 		Keys: bson.D{
-			{"bk_module_id", int32(1)},
-			{"bk_biz_id", int32(1)},
+			{"bk_module_id", 1},
+			{"bk_biz_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "idx_unique_moduleID",
 		Keys: bson.D{{
-			"bk_module_id", int32(1)},
+			"bk_module_id", 1},
 		},
 		Unique:     true,
 		Background: true,
@@ -94,22 +94,22 @@ var deprecatedModuleBaseIndexes = []types.Index{
 	{
 		Name: "bk_idx_set_template_id_service_template_id",
 		Keys: bson.D{
-			{"set_template_id", int32(1)},
-			{"service_template_id", int32(1)},
+			{"set_template_id", 1},
+			{"service_template_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_idx_set_template_id",
 		Keys: bson.D{{
-			"set_template_id", int32(1)},
+			"set_template_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_idx_service_template_id",
 		Keys: bson.D{{
-			"service_template_id", int32(1)},
+			"service_template_id", 1},
 		},
 		Background: true,
 	},
