+++ src/common/index/collections/modulehostconfig.go
@@ -32,7 +32,7 @@ func init() {
 var commModuleHostConfigIndexes = []types.Index{
 	{
 		Name:       "bkcc_idx_bkBizID_bkHostID",
-		Keys:       bson.D{{"bk_biz_id", int32(1)}, {"bk_host_id", int32(1)}},
+		Keys:       bson.D{{"bk_biz_id", 1}, {"bk_host_id", 1}},
 		Background: true,
 	},
 }
@@ -42,52 +42,52 @@ var deprecatedModuleHostConfigIndexes = []types.Index{
 	{
 		Name: "bk_biz_id_1",
 		Keys: bson.D{{
-			"bk_biz_id", int32(1)},
+			"bk_biz_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_host_id_1",
 		Keys: bson.D{{
-			"bk_host_id", int32(1)},
+			"bk_host_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_module_id_1",
 		Keys: bson.D{{
-			"bk_module_id", int32(1)},
+			"bk_module_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_set_id_1",
 		Keys: bson.D{{
-			"bk_set_id", int32(1)},
+			"bk_set_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_module_id_1_bk_biz_id_1",
 		Keys: bson.D{
-			{"bk_module_id", int32(1)},
-			{"bk_biz_id", int32(1)},
+			{"bk_module_id", 1},
+			{"bk_biz_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_set_id_1_bk_biz_id_1",
 		Keys: bson.D{
-			{"bk_set_id", int32(1)},
-			{"bk_biz_id", int32(1)},
+			{"bk_set_id", 1},
+			{"bk_biz_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "idx_unique_moduleID_hostID",
 		Keys: bson.D{
-			{"bk_module_id", int32(1)},
-			{"bk_host_id", int32(1)},
+			{"bk_module_id", 1},
+			{"bk_host_id", 1},
 		},
 		Unique:     true,
 		Background: true,
