+++ src/common/index/collections/netcollectdevice.go
@@ -36,21 +36,21 @@ var deprecatedNetcollectDeviceIndexes = []types.Index{
 	{
 		Name: "device_id_1",
 		Keys: bson.D{{
-			"device_id", int32(1)},
+			"device_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "device_name_1",
 		Keys: bson.D{{
-			"device_name", int32(1)},
+			"device_name", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_supplier_account_1",
 		Keys: bson.D{{
-			"bk_supplier_account", int32(1)},
+			"bk_supplier_account", 1},
 		},
 		Background: true,
 	},
