+++ src/common/index/collections/objasst.go
@@ -35,37 +35,37 @@ var deprecatedObjAsstIndexes = []types.Index{
 	{
 		Name: "bk_obj_id_1",
 		Keys: bson.D{{
-			"bk_obj_id", int32(1)},
+			"bk_obj_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_asst_obj_id_1",
 		Keys: bson.D{{
-			"bk_asst_obj_id", int32(1)},
+			"bk_asst_obj_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_supplier_account_1",
 		Keys: bson.D{{
-			"bk_supplier_account", int32(1)},
+			"bk_supplier_account", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "bk_obj_id_1_bk_asst_obj_id_1_bk_asst_id_1",
 		Keys: bson.D{
-			{"bk_obj_id", int32(1)},
-			{"bk_asst_obj_id", int32(1)},
-			{"bk_asst_id", int32(1)},
+			{"bk_obj_id", 1},
+			{"bk_asst_obj_id", 1},
+			{"bk_asst_id", 1},
 		},
 		Background: true,
 	},
 	{
 		Name: "idx_unique_id",
 		Keys: bson.D{{
-			"id", int32(1)},
+			"id", 1},
 		},
 		Unique:     true,
 		Background: true,
